requirements.txt
asyncio==3.4.3
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
python-dotenv==1.0.0
requests==2.28.2
resilient-caller==0.2.1
soupsieve==2.4
tls-client==0.2
urllib3==1.26.15
